convert component from class component to functional component
This component is being converted into a functional component. The component does
not have state, however, it is consuming props from PoliticsMap. The state in
the parent component should likely have its own context to store state, because
this component will be broken down into several smaller components to follow the
single responsibility principle. We also make use of useRef in this component,
react-flippy requires us to have a reference to the DOM to perform the animation
and re-render accordingly to ensure that its not receiving the wrong DOM node because
of a stale closure.